Backend Implementation
This is the first update of the backend Implementation.
I have created the main method which implements the use of fastapi and the root.
install_requests: implemented the SQL queries for sending requests to the admin when an installation request on the home station is requested. This includes get and post requests.
request.py: Creates the object of the request. This includes the status, base (information, inclduing name size and source, etc.), install request read.
database.py: Implemented sqlalchemy and using sqlite to create a database URL that interacts with the requests.